Remove the requirement regarding HTTP status code 406
Rework the clause on multilinguality and annex D.

In addition: remove headings that say "INSPIRE-specific requirements",
as all the requirements are per definition INSPIRE-specific, and it
may confuse as the principle is to limit INSPIRE-specific extensions
to a minimum.

In addition: editorial changes:
edit some references to use Reference-style Links, fix typos.